process_command_svn(): add partial parsing of length:string pairs
this is in preparation for adding support for svn log; as commit
messages may contain unbalanced pairs of parens that could confuse
the logic here, which is used to determine how many response bytes
to read, and at the same time to partition data blocks into groups.

the added code isn't bullet-proof, there are a number of cases that
would require complex logic to handle them involving the received
packet data cut off during read of the length, or during the string
itself. in such a case we fall back to the old logic of parens
counting.
this should not pose a big problem in practice as commit messages
are usually the only "unsanitized" input and typically shorter
than a single packet. the only problematic case could be a huge
commit message with or without unbalanced parens.
if we hit such a problem in the future, we might fall back to a
different strategy, such as reading byte by byte only for log
entries.