implement log and info
this should be sufficient to implement a basic svn2git client.
log and info both work on either an URL or a local directory,
however using a local directory only information about the
last checked out revision can be retrieved.

changed filenames of svnup savedata to .svnup/revision and
.svnup/known_files.

there's a rather big block where i added an if statement around
and didn't indent yet, in order not to spam the commit diff
with huge whitespace noise.