fix failure detection in svn log scenario
for some reason, the svn log result is presented in a different
format than all other svn responses we process, which confused
the reading logic. also one group less is returned in case of
this error, so by expecting 3 groups the code would try to read
on even if we already have the full response and block until
remote disconnect happens.