replace super-confusing buffer*** code
svnup uses a clever technique: it sends multiple SVN/HTTP requests
in a single write() transaction (the HTTP requests use Connection:
keep-alive for that) which causes the kernel to buffer up the
incoming responses to each request, but this comes at quite some
cost in added complexity of book-keeping.
for example, in order to split out the actual contents of the
received files, the length of the files need to be gathered in
advance.

in my opinion it's questionable whether this approach truly
results in a faster transfer (at least in the HTTP case), as
the time to enter/leave the kernel is maybe 1000 cycles vs network
latency in millisecond range; and a single send-and-recv cycle
per file might be just as fast as long as the connection is
kept alive.

in order to stuff all requests into big buffers (but not too big,
the HTTP code used a max of 95 requests per write syscall, and the
SVN code roughly 30KB) a complicated ad-hoc implementation of
string buffers was used.

i replaced those hard-to-decipher buffer handling routines with
a simple stringlist implementation (taken from my libulz project),
and a function that merges a set of strings (here: requests)
into a single big string. this makes the code much more readable,
and also more performant as we don't need to strcat() to a 30KB
big buffer which results in up to 30.000 cycles scanning for the
string end for each filename addition.