if available, use inline-props for HTTP checkouts
if inline-props are available and requested, most of the additional
attributes (except filesize) are transmitted as part of the initial
report request, so we don't have to file an additional request per
file checked into the repo, at the cost of a larger report.
it is hoped that this should speed up HTTP requests noticeably.