replace md5 check of all local files with a known_files lookup
even though having all md5s already stored in known_files file and
the associated RB tree, the client wouldn't actually use it any-
where and instead check the md5 against the files on the filesystem.
with a big number of files, or especially big files, this can take
a huge amount of time, as it needs to be done for every single file
in the repo.
additionally, the now deleted confirm_md5 function would slurp each
file entirely into memory to take its checksum, rather than reading
it blockwise, which could easily cause an OOM kill on a big file.
therefore it has been a thorn in my eye for quite some time and
on my todo-list to overhaul; but now it turned out it isn't needed
at all.